Add delete confirmation popup for delete character.

+++ src/app/components/characters/characters-list/character-item/character-item-list.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
+import { ConfirmationService } from 'primeng/api';
 import Character from 'src/app/model/character.model';
 import { removeCharacter } from 'src/app/state/character/character.action';
 import { AppState } from 'src/app/state/state';
@@ -12,11 +13,23 @@ import { AppState } from 'src/app/state/state';
 export class CharacterItemListComponent implements OnInit {
     @Input() character: Character;
 
-    constructor(private store: Store<AppState>) {}
-
+    constructor(
+        private store: Store<AppState>,
+        private confirmationService: ConfirmationService
+    ) {}
     ngOnInit(): void {}
 
-    remove(character: Character) {
-        this.store.dispatch(removeCharacter({ character }));
+    confirmDelete(event: Event, character: Character) {
+        this.confirmationService.confirm({
+            target: event.target,
+            message: 'Are you sure that you want to proceed?',
+            icon: 'pi pi-exclamation-triangle',
+            accept: () => {
+                this.store.dispatch(removeCharacter({ character }));
+            },
+            reject: () => {
+                //reject action
+            },
+        });
     }
 }
